adapt scripts to handle empty annotations

+++ scripts/exonerate.sh
@@ -24,13 +24,12 @@ function get_help() {
 
 		echo "USAGE(S):"
 		echo -e "\
-	\t$PROGRAM [-a <address>] [-f <TSV file>] [-t <target FASTA file>] -o <output directory> <query FASTA file>\n \
+	\t$PROGRAM [-a <address>] [-t <target FASTA file>] -o <output directory> <query FASTA file> <annotation TSV file>\n \
 	" | table
 
 		echo "OPTION(S):"
 		echo -e "\
 	\t-a <address>\temail address for alerts\n \
-	\t-f <TSV>\tannotation TSV file\n \
 	\t-h\tshow this help menu\n \
 	\t-o <directory>\toutput directory\t(required)\n \
 	\t-t <FASTA>\ttarget FASTA file\t(default = \$ROOT_DIR/amp_seqs/amps.\$CLASS.prot.combined.faa)\n \
@@ -72,23 +71,24 @@ fi
 target=""
 outdir=""
 email=false
-file=""
 # 4 - getopts
-while getopts :a:f:ho:t: opt; do
+while getopts :a:ho:t: opt; do
 	case $opt in
 	a)
 		address="$OPTARG"
 		email=true
 		;;
-	f) file="$(realpath $OPTARG)" ;;
 	h) get_help ;;
 	o) outdir=$(realpath $OPTARG) ;;
-	t) target=$(realpath $OPTARG) ;;
-	\?) print_error "Invalid option: -$OPTARG" ;;
+	t) target=$(realpath $OPTARG) ;; \?) print_error "Invalid option: -$OPTARG" ;;
 	esac
 done
 
 shift $((OPTIND - 1))
+
+if [[ "$#" -ne 2 ]]; then
+	print_error "Incorrect number of arguments."
+fi
 
 # 6 - check input files
 if [[ -z $outdir ]]; then
@@ -103,12 +103,10 @@ elif [[ ! -s $(realpath $1) ]]; then
 	print_error "Input file $(realpath $1) is empty."
 fi
 
-if [[ -n $file ]]; then
-	if [[ ! -f $file ]]; then
-		print_error "TSV file $file does not exist."
-	elif [[ ! -s $file ]]; then
-		print_error "TSV file $file is empty."
-	fi
+if [[ ! -f $(realpath $2) ]]; then
+	print_error "Input file $(realpath $2) does not exist."
+elif [[ ! -s $(realpath $2) ]]; then
+	print_error "Input file $(realpath $2) is empty."
 fi
 
 # if workdir not set, infer from indir
@@ -134,6 +132,7 @@ fi
 if [[ -z "$target" ]]; then
 	target=$ROOT_DIR/amp_seqs/amps.${class^}.prot.combined.faa
 fi
+
 # 8 - print env details
 {
 	echo "HOSTNAME: $(hostname)"
@@ -145,48 +144,77 @@ fi
 } 1>&2
 
 query=$(realpath $1)
-
+file=$(realpath $2)
 echo "PROGRAM: $(command -v $RUN_EXONERATE)" 1>&2
 echo -e "VERSION: $($RUN_EXONERATE --version 2>&1 | head -n1 | awk '{print $NF}')\n" 1>&2
 
 echo "Running Exonerate..." 1>&2
-echo -e "COMMAND: $RUN_EXONERATE --query $query --target $target --querytype protein --targettype protein --ryo "Summary: %qi %ti %pi\n" --showvulgar false >$outdir/amps.exonerate.out\n" 1>&2
+echo -e "COMMAND: $RUN_EXONERATE --query $query --target $target --querytype protein --targettype protein --ryo \"Summary: %qi %ti %pi\\\n\" --showvulgar false >$outdir/amps.exonerate.out\n" 1>&2
 $RUN_EXONERATE --query $query --target $target --querytype protein --targettype protein --ryo "Summary: %qi %ti %pi\n" --showvulgar false >$outdir/amps.exonerate.out
+if [[ "$(wc -l $outdir/amps.exonerate.out | awk '{print $1}')" -gt 3 ]]; then
+	# there are known AMPs!
+	echo "Extracting summary..." 1>&2
+	echo -e "COMMAND: grep '^Summary:' $outdir/amps.exonerate.out | cut -d' ' -f2- $outdir/amps.exonerate.summary.out\n" 1>&2
+	grep '^Summary:' $outdir/amps.exonerate.out | cut -d' ' -f2- >$outdir/amps.exonerate.summary.out
 
-echo "Extracting summary..." 1>&2
-echo -e "COMMAND: grep '^Summary:' $outdir/amps.exonerate.out | cut -d' ' -f2- $outdir/amps.exonerate.summary.out\n" 1>&2
-grep '^Summary:' $outdir/amps.exonerate.out | cut -d' ' -f2- >$outdir/amps.exonerate.summary.out
+	echo "Filtering for known AMPs..." 1>&2
+	echo -e "COMMAND: $RUN_SEQTK subseq $query <(awk '{if(\$3==100) print \$1}' $outdir/amps.exonerate.summary.out | sort -u) >$outdir/known.amps.exonerate.nr.faa\n" 1>&2
+	$RUN_SEQTK subseq $query <(awk '{if($3==100) print $1}' $outdir/amps.exonerate.summary.out | sort -u) >$outdir/known.amps.exonerate.nr.faa
 
-echo "Filtering for known AMPs..." 1>&2
-echo -e "COMMAND: $RUN_SEQTK subseq $query <(awk '{print \$1}' $outdir/amps.exonerate.summary.out | sort -u) >$outdir/known.amps.exonerate.nr.faa\n" 1>&2
-$RUN_SEQTK subseq $query <(awk '{print $1}' $outdir/amps.exonerate.summary.out | sort -u) >$outdir/known.amps.exonerate.nr.faa
+	echo "Labelling known AMPs..." 1>&2
+	# echo -e "COMMAND: sed -i '/^>/ s/ length=/_known&/' $outdir/known.amps.exonerate.nr.faa" 1>&2
+	sed -i '/^>/ s/ length=/_known&/' $outdir/known.amps.exonerate.nr.faa
 
-echo "Labelling known AMPs..." 1>&2
-# echo -e "COMMAND: sed -i '/^>/ s/ length=/_known&/' $outdir/known.amps.exonerate.nr.faa" 1>&2
-sed -i '/^>/ s/ length=/_known&/' $outdir/known.amps.exonerate.nr.faa
+	# label with known AMP accession
+	while IFS=' ' read novel known; do
+		sed -i "/${novel}_known/ s/$/ exonerate=$known/" $outdir/known.amps.exonerate.nr.faa
+	done < <(awk '{if($3==100) print $1, $2}' $outdir/amps.exonerate.summary.out | sort -u)
 
-# label with known AMP accession
-while IFS=' ' read novel known; do
-	sed -i "/${novel}_known/ s/$/ known=$known/" $outdir/known.amps.exonerate.nr.faa
-done < <(awk '{print $1, $2}' $outdir/amps.exonerate.summary.out | sort -u)
+	echo "Filtering for novel AMPs..." 1>&2
+	echo -e "COMMAND: $RUN_SEQTK subseq $query <(grep -vFf <(awk '{if(\$3==100) print \$1}' $outdir/amps.exonerate.summary.out | sort -u) <(awk '/^>/ {print \$1}' $query | tr -d '>') | sort -u) >$outdir/novel.amps.exonerate.nr.faa\n" 1>&2
+	$RUN_SEQTK subseq $query <(grep -vFf <(awk '{if($3==100) print $1}' $outdir/amps.exonerate.summary.out | sort -u) <(awk '/^>/ {print $1}' $query | tr -d '>') | sort -u) >$outdir/novel.amps.exonerate.nr.faa
 
-echo "Filtering for novel AMPs..." 1>&2
-echo -e "COMMAND: $RUN_SEQTK subseq $query <(grep -vFf <(awk '{print \$1}' $outdir/amps.exonerate.summary.out | sort -u) <(grep '^>' $query | tr -d '>' | sort -u)) >$outdir/novel.amps.exonerate.nr.faa\n" 1>&2
-$RUN_SEQTK subseq $query <(grep -vFf <(awk '{print $1}' $outdir/amps.exonerate.summary.out | sort -u) <(grep '^>' $query | tr -d '>')) >$outdir/novel.amps.exonerate.nr.faa
+	# echo -e "COMMAND: $RUN_SEQTK subseq $query <(grep -vFf <(awk '{print \$1}' $outdir/amps.exonerate.summary.out | sort -u) <(grep '^>' $query | tr -d '>' | sort -u)) >$outdir/novel.amps.exonerate.nr.faa\n" 1>&2
+	# $RUN_SEQTK subseq $query <(grep -vFf <(awk '{print $1}' $outdir/amps.exonerate.summary.out | sort -u) <(grep '^>' $query | tr -d '>')) >$outdir/novel.amps.exonerate.nr.faa
 
-echo "Labelling novel AMPs..." 1>&2
-echo -e "COMMAND: sed -i '/^>/ s/ length=/_novel&/' $outdir/novel.amps.exonerate.nr.faa\n" 1>&2
-sed -i '/^>/ s/ length=/_novel&/' $outdir/novel.amps.exonerate.nr.faa
+	echo "Labelling novel AMPs..." 1>&2
+	echo -e "COMMAND: sed -i '/^>/ s/ length=/_novel&/' $outdir/novel.amps.exonerate.nr.faa\n" 1>&2
+	sed -i '/^>/ s/ length=/_novel&/' $outdir/novel.amps.exonerate.nr.faa
 
-echo "Combining the two files..." 1>&2
-echo -e "COMMAND: cat $outdir/known.amps.exonerate.nr.faa $outdir/novel.amps.exonerate.nr.faa >$outdir/labelled.amps.exonerate.nr.faa\n" 1>&2
-cat $outdir/known.amps.exonerate.nr.faa $outdir/novel.amps.exonerate.nr.faa >$outdir/labelled.amps.exonerate.nr.faa
+	echo "Combining the two files..." 1>&2
+	echo -e "COMMAND: cat $outdir/known.amps.exonerate.nr.faa $outdir/novel.amps.exonerate.nr.faa >$outdir/labelled.amps.exonerate.nr.faa\n" 1>&2
+	cat $outdir/known.amps.exonerate.nr.faa $outdir/novel.amps.exonerate.nr.faa >$outdir/labelled.amps.exonerate.nr.faa
+
+	# add the label to the annotation TSV as well
+	for seq in $(grep '_novel' $outdir/labelled.amps.exonerate.nr.faa | tr -d '>' | sed 's/_novel//'); do
+		sed -i "s/${seq}\t/${seq}_novel\t/" $file
+	done
+
+	for seq in $(grep '_known' $outdir/labelled.amps.exonerate.nr.faa | tr -d '>' | sed 's/_known//'); do
+		sed -i "s/${seq}\t/${seq}_known\t/" $file
+	done
+else
+	# no known AMPs
+	echo -e "No alignments detected--there are no known AMPs. All AMPs are novel!\n" 1>&2
+
+	cp $query $outdir/novel.amps.exonerate.nr.faa
+	touch $outdir/known.amps.exonerate.nr.faa
+
+	echo "Labelling novel AMPs..." 1>&2
+	sed -i '/^>/ s/ length=/_novel&/' $outdir/novel.amps.exonerate.nr.faa
+	cat $outdir/known.amps.exonerate.nr.faa $outdir/novel.amps.exonerate.nr.faa >$outdir/labelled.amps.exonerate.nr.faa
+fi
+
+num_novel=$(grep -c '_novel' $outdir/labelled.amps.exonerate.nr.faa || true)
+num_total=$(grep -c '^>' $outdir/labelled.amps.exonerate.nr.faa)
+
+echo -e "Number of Novel AMPs: $(printf "%'d" $num_novel)/$(printf "%'d" $num_total)\n" 1>&2
 
 if [[ -n $file ]]; then
-	echo "Filtering annotations for novel AMPs..." 1>&2
-	head -n1 $file >$outdir/annotated.novel.amps.tsv
-	echo -e "COMMAND: grep -Ff <(awk '{print $1}' $outdir/amps.exonerate.summary.out) $file >>$outdir/annotated.novel.amps.tsv" 1>&2
-	grep -Ff <(awk '{print $1}' $outdir/amps.exonerate.summary.out) $file >>$outdir/annotated.novel.amps.tsv
+	echo -e "Output(s): $outdir/labelled.amps.exonerate.nr.faa\n $file\n \
+	" | column -s ' ' -t 1>&2
+else
+	echo -e "Output: $outdir/novel.amps.exonerate.nr.faa\n" 1>&2
 fi
 
 default_name="$(realpath -s $(dirname $outdir)/exonerate)"
@@ -210,18 +238,11 @@ if [[ "$default_name" != "$outdir" ]]; then
 		(cd $(dirname $outdir) && ln -fs $(basename $outdir) $(basename $default_name))
 	fi
 fi
-echo -e "END: $(date)\n" 1>&2
+echo -e "\nEND: $(date)\n" 1>&2
 # echo 1>&2
 
 echo -e "STATUS: DONE.\n" 1>&2
 touch $outdir/EXONERATE.DONE
-
-if [[ -n $file ]]; then
-	echo -e "Output(s): $outdir/novel.amps.exonerate.nr.faa\n $outdir/annotated.novel.amps.tsv\
-	" | column -s ' ' -t 1>&2
-else
-	echo "Output: $outdir/novel.amps.exonerate.nr.faa" 1>&2
-fi
 
 if [[ "$email" = true ]]; then
 	# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
